requirements.txt
bitmap==0.0.7
certifi==2020.4.5.1
chardet==3.0.4
configparser==5.0.0
cycler==0.10.0
dicom2nifti==2.2.10
future==0.18.2
idna==2.9
imageio==2.8.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
nibabel==3.1.0
numpy==1.18.4
packaging==20.4
peewee==3.13.3
Pillow==7.1.2
psycopg2==2.8.5
pydicom==1.4.2
pyorthanc==0.2.12
pyparsing==2.4.7
pyside-material==1.11
PySide2==5.14.0
python-dateutil==2.8.1
requests==2.23.0
scipy==1.4.1
shiboken2==5.14.0
SimpleITK==1.2.4
six==1.14.0
SQLAlchemy==1.3.16
urllib3==1.25.9
